+++ code/server/server.h
@@ -103,7 +103,7 @@ typedef struct {
 										// order, otherwise the delta compression will fail
 	int				messageSent;		// time the message was transmitted
 	int				messageAcked;		// time the message was acked
-	int				messageSize;		// used to rate drop packets
+	size_t			messageSize;		// used to rate drop packets
 } clientSnapshot_t;
 
 typedef enum {
@@ -147,13 +147,13 @@ typedef struct client_s {
 	// downloading
 	char			downloadName[MAX_QPATH]; // if not empty string, we are downloading
 	fileHandle_t	download;			// file being downloaded
- 	int				downloadSize;		// total bytes (can't use EOF because of paks)
- 	int				downloadCount;		// bytes sent
+	size_t			downloadSize;		// total bytes (can't use EOF because of paks)
+	size_t			downloadCount;		// bytes sent
 	int				downloadClientBlock;	// last block we sent to the client, awaiting ack
 	int				downloadCurrentBlock;	// current block number
 	int				downloadXmitBlock;	// last block we xmited
 	unsigned char	*downloadBlocks[MAX_DOWNLOAD_WINDOW];	// the buffers for the download blocks
-	int				downloadBlockSize[MAX_DOWNLOAD_WINDOW];
+	size_t			downloadBlockSize[MAX_DOWNLOAD_WINDOW];
 	qboolean		downloadEOF;		// We have sent the EOF block
 	int				downloadSendTime;	// time we last got an ack from the client
